+++ Commands/Standard/SVG.animate.ps1
@@ -1509,86 +1509,8 @@ function SVG.animate {
 .Example
     svg -Content @(
         svg.polygon -Points "25,50 50,75 75,50 50,25" -Fill '#4488ff' @(
-            svg.animate -AttributeName points -to "0,0 0,100 100,100, 100,0" -Dur 2s -Id morph1 -Begin '0s;morph2.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '0' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '1' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
+            svg.animate -AttributeName points -Values "25,50 50,75 75,50 50,25;0,0 0,100 100,100, 100,0; 25,50 50,75 75,50 50,25" -Dur 2s -Id morph1 -RepeatCount 'indefinite' -AttributeType XML
         )
-        svg.polygon -Points "0,0 0,100 100,100, 100,0" -Fill '#4488ff' @(
-            svg.animate -AttributeName opacity -Values '1' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName points -to "25,50 50,75 75,50 50,25" -Dur 2s -Id morph2 -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '0' -Dur '0.0s' -Begin 'morph2.end' -AttributeType XML
-        ) -Opacity 0
-    
-    ) -ViewBox 100,100
-.Example
-    svg -Content @(
-        svg.polygon -Points "25,50 50,75 75,50 50,25" -Fill '#4488ff' @(
-            svg.animate -AttributeName points -to "0,0 0,100 100,100, 100,0" -Dur 2s -Id morph1 -Begin '0s;morph2.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '0' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '1' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-        )
-        svg.polygon -Points "0,0 0,100 100,100, 100,0" -Fill '#4488ff' @(
-            svg.animate -AttributeName opacity -Values '1' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName points -to "25,50 50,75 75,50 50,25" -Dur 2s -Id morph2 -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '0' -Dur '0.0s' -Begin 'morph2.end' -AttributeType XML
-        ) -Opacity 0
-    
-    ) -ViewBox 100,100
-.Example
-    svg -Content @(
-        svg.polygon -Points "25,50 50,75 75,50 50,25" -Fill '#4488ff' @(
-            svg.animate -AttributeName points -to "0,0 0,100 100,100, 100,0" -Dur 2s -Id morph1 -Begin '0s;morph2.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '0' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '1' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-        )
-        svg.polygon -Points "0,0 0,100 100,100, 100,0" -Fill '#4488ff' @(
-            svg.animate -AttributeName opacity -Values '1' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName points -to "25,50 50,75 75,50 50,25" -Dur 2s -Id morph2 -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '0' -Dur '0.0s' -Begin 'morph2.end' -AttributeType XML
-        ) -Opacity 0
-    
-    ) -ViewBox 100,100
-.Example
-    svg -Content @(
-        svg.polygon -Points "25,50 50,75 75,50 50,25" -Fill '#4488ff' @(
-            svg.animate -AttributeName points -to "0,0 0,100 100,100, 100,0" -Dur 2s -Id morph1 -Begin '0s;morph2.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '0' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '1' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-        )
-        svg.polygon -Points "0,0 0,100 100,100, 100,0" -Fill '#4488ff' @(
-            svg.animate -AttributeName opacity -Values '1' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName points -to "25,50 50,75 75,50 50,25" -Dur 2s -Id morph2 -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '0' -Dur '0.0s' -Begin 'morph2.end' -AttributeType XML
-        ) -Opacity 0
-    
-    ) -ViewBox 100,100
-.Example
-    svg -Content @(
-        svg.polygon -Points "25,50 50,75 75,50 50,25" -Fill '#4488ff' @(
-            svg.animate -AttributeName points -to "0,0 0,100 100,100, 100,0" -Dur 2s -Id morph1 -Begin '0s;morph2.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '0' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '1' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-        )
-        svg.polygon -Points "0,0 0,100 100,100, 100,0" -Fill '#4488ff' @(
-            svg.animate -AttributeName opacity -Values '1' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName points -to "25,50 50,75 75,50 50,25" -Dur 2s -Id morph2 -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '0' -Dur '0.0s' -Begin 'morph2.end' -AttributeType XML
-        ) -Opacity 0
-    
-    ) -ViewBox 100,100
-.Example
-    svg -Content @(
-        svg.polygon -Points "25,50 50,75 75,50 50,25" -Fill '#4488ff' @(
-            svg.animate -AttributeName points -to "0,0 0,100 100,100, 100,0" -Dur 2s -Id morph1 -Begin '0s;morph2.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '0' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '1' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-        )
-        svg.polygon -Points "0,0 0,100 100,100, 100,0" -Fill '#4488ff' @(
-            svg.animate -AttributeName opacity -Values '1' -Dur '0.0s' -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName points -to "25,50 50,75 75,50 50,25" -Dur 2s -Id morph2 -Begin 'morph1.end' -AttributeType XML
-            svg.animate -AttributeName opacity -Values '0' -Dur '0.0s' -Begin 'morph2.end' -AttributeType XML
-        ) -Opacity 0
-    
     ) -ViewBox 100,100
 .Example
     svg -Content @(
@@ -1692,14 +1614,14 @@ function SVG.animate {
     [float]$CenterY  = 100,
     # The radius coordinate of the shape.  This will decrease by 1/RepeatCount each time.
     [float]$Radius   = 100,
-    # The number of sides
-    $SideCount  = 3,
+    # The number of sides.  A file will be generated for each unique value provided.
+    [int[]]$SideCount  = 3..6,
     # The total rotation of the innermost element,
     $TotalRotation  = 180,
     # The total duration of any animations.
     [timespan]$duration = '00:00:03.75',
     # A palette of colors to alternate thru
-    [string[]]$colors = @('#112244','#224488',"#4488ff"),
+    [string[]]$Color = @('#112244','#224488',"#4488ff"),
     # The type of the shape. (either Star or ConvexPolygon)
     [ValidateSet("Star", "ConvexPolygon")]
     [string]
@@ -1710,36 +1632,37 @@ function SVG.animate {
     $AnimateOpacity
     )
     
-    $Splat = [Ordered]@{
-        SideCount = $SideCount
-        Fill = 'transparent'
-        CenterX = $CenterX
-        CenterY = $CenterY
-    }
     
     $shapeCommand = $ExecutionContext.SessionState.InvokeCommand.GetCommand("SVG.$ShapeType", "Function")
-    SVG -ViewBox (($CenterX * 2), ($CenterY * 2)) @(
-        SVG.rect -Width 1000% -Height 1000% -X -500% -Y -500% -Fill 'black'
     
-        0..($RepeatCount -1) |
-            . $shapeCommand @Splat -Rotate {
-                    $_ * ($totalRotation / $RepeatCount)
-            } -Radius {
-                $Radius - (
-                    $_ * ($Radius / $RepeatCount)
-                )
-            } -Stroke {
-                $colors[$_ % $colors.Length]
-            } -Children {
-                $toRotation =  $(360 * ([Math]::Ceiling(($_ + 1)/10)))
-                SVG.animateTransform -From "0 $centerX $centerY" -To "$toRotation $centerX $centerY" -Dur $duration -AttributeName transform -Type 'rotate' -RepeatCount 'indefinite'
-                $lowOpacity = [double]($_)/$RepeatCount
-                $highOpacity = 1.0 - [double]($_)/$RepeatCount
-                if ($AnimateOpacity) {
-                    SVG.animate -AttributeName opacity -Values "$highOpacity;$lowOpacity;$highOpacity" -Dur $dur -RepeatCount 'indefinite'
+    foreach ($Sides in $SideCount) {
+        $Splat = [Ordered]@{
+            SideCount = $Sides
+            Fill = 'transparent'
+            CenterX = $CenterX
+            CenterY = $CenterY
+        }
+    
+        SVG -ViewBox (($CenterX * 2), ($CenterY * 2)) @(
+            0..($RepeatCount -1) |
+                & $shapeCommand @Splat -Rotate {
+                        $_ * ($totalRotation / $RepeatCount)
+                } -Radius {
+                    $Radius - (
+                        $_ * ($Radius / $RepeatCount)
+                    )
+                } -Stroke {
+                    $Color[$_ % $color.Length]
+                } -Children {
+                    $toRotation =  $(360 * ([Math]::Ceiling(($_ + 1)/10)))
+                    SVG.animateTransform -From "0 $centerX $centerY" -To "$toRotation $centerX $centerY" -Dur $duration -AttributeName transform -Type 'rotate' -RepeatCount 'indefinite'
+                    $lowOpacity = [double]($_)/$RepeatCount
+                    $highOpacity = 1.0 - [double]($_)/$RepeatCount
+                    if ($AnimateOpacity) {
+                        SVG.animate -AttributeName opacity -Values "$highOpacity;$lowOpacity;$highOpacity" -Dur $dur -RepeatCount 'indefinite'
+                    }
                 }
-            }
-    )
+        )
 .Example
     svg @(
         svg.ConvexPolygon -SideCount 8 -Rotate (360/16) -Fill '#dd0000' -Stroke white -CenterX 100 -CenterY 100 -Radius 100
@@ -1799,6 +1722,11 @@ function SVG.animate {
             svg.tspan -Content "GIFS" -Id gif
             svg.animate -Values '28;30;28' -Dur 5s -AttributeName font-size -RepeatDur 'indefinite'
         )
+    ) -ViewBox 200,200
+.Example
+    svg @(
+        svg.ConvexPolygon -SideCount 8 -Rotate (360/16) -Fill '#4488ff' -Stroke '#4488ff' -CenterX 100 -CenterY 100 -Radius 50
+        SVG.animate -AttributeName viewBox -Values "0 0 200 200; 50 50 100 100; 0 0 200 200" -RepeatCount 'indefinite' -Dur 3.9
     ) -ViewBox 200,200
 .Link
     https://pssvg.start-automating.com/SVG.animate