+++ Commands/Standard/SVG.pattern.ps1
@@ -104,6 +104,18 @@ function SVG.pattern {
         SVG.rect -Width 800 -Height 800 -Fill "url(#fillPattern)" -X 0 -Y 0
     )
 .Example
+    SVG -Content @(
+        SVG.defs -Content @(
+    
+            SVG.pattern -id 'HexagonPattern' -patternUnits 'userSpaceOnUse' -width '174' -height '200' -patternTransform 'scale(.5)' -Content @(
+                SVG.Hexagon -CenterX 87 -CenterY 100 -Radius 100 -Fill transparent -Stroke '#4488ff' -Comment "A hexagon with a" -Class 'foreground-fill'
+            ) -Comment "A faint hexagon Pattern, rendered 4x at a base scale of 174 by 200"
+        )
+        $hugeSize = 20000
+    
+        SVG.rect -width "$hugeSize%" -height "$hugeSize%" -fill 'url(#HexagonPattern)'
+    )
+.Example
     svg @(
         svg.defs @(
             svg.pattern -Id 'SimplePattern' -Width .1 -Height .1 -Content @(